FIX: Make channel ids wide enough for extended CAN frames

+++ src/sdk/DataFrame.h
@@ -13,7 +13,7 @@ enum class FrameDirection {
     Transmit
 };
 
-using ChannelId = quint32;
+using ChannelId = quint64;
 
 struct ChannelSample {
     ChannelId index = 0;
